+++ lib/solar_test/src/reporters/xml_reporter.cpp
@@ -195,7 +195,7 @@ void XmlReporter::write_system_err_element(const TestResult& result) {
 
 std::string XmlReporter::xml_escape(const std::string& text) const {
   std::string escaped;
-  escaped.reserve(text.length() * 1.2);  // Reserve some extra space
+  escaped.reserve(static_cast<size_t>(text.length() * 1.2));  // Reserve some extra space
 
   for (char c : text) {
     switch (c) {
